finalized salary endpoint

+++ src/main/java/com/cs322/ors/model/Salary.java
@@ -1,9 +1,15 @@
 package com.cs322.ors.model;
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToOne;
+
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.JsonIdentityReference;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 
 @Entity
 public class Salary {
@@ -11,16 +17,18 @@ public class Salary {
 @GeneratedValue(strategy = GenerationType.IDENTITY)
 private long id;
 
+//@JsonIdentityInfo(property = "id", generator = ObjectIdGenerators.PropertyGenerator.class)
+//@JsonIdentityReference(alwaysAsId = true)
 @OneToOne(optional = false)
 private User user;
 
-private long amount;
+private BigDecimal amount;
 private String description;
 
 public Salary() {
 }
 
-public Salary(User user, long amount, String description) {
+public Salary(User user, BigDecimal amount, String description) {
     this.user = user;
     this.amount = amount;
     this.description = description;
@@ -43,11 +51,11 @@ public void setUser(User user) {
     this.user = user;
 }
 
-public long getAmount() {
+public BigDecimal getAmount() {
     return amount;
 }
 
-public void setAmount(long amount) {
+public void setAmount(BigDecimal amount) {
     this.amount = amount;
 }
